+++ packages/ts-newskit/src/theme/times-web-light/style-presets/index.ts
@@ -9,6 +9,8 @@ import { pageStylePresets } from './page';
 import { secondaryNavigationStylePresets } from './secondary-nav';
 import { sliceStylePresets } from './slices';
 import { textStylePresets } from './text';
+import { flagStylePresets } from './flag';
+import { accordionStylePresets } from './accordion';
 
 export const stylePresets = {
   ...articleListStylePresets,
@@ -21,5 +23,7 @@ export const stylePresets = {
   ...navStylePresets,
   ...secondaryNavigationStylePresets,
   ...leadStoryStylePresets,
-  ...sliceStylePresets
+  ...sliceStylePresets,
+  ...flagStylePresets,
+  ...accordionStylePresets
 };